host_requirements.txt
invoke==2.0.0
python-box==7.0.0
PyYAML==6.0
